fix analysis
