+++ lib/built_redux.dart
@@ -4,4 +4,5 @@ export 'src/action.dart';
 export 'src/built_reducer.dart';
 export 'src/middleware.dart';
 export 'src/store.dart';
+export 'src/store_change.dart';
 export 'src/typedefs.dart';
